+++ csharp-solution/SpeechFlowCsharp/AudioProcessing/IVoiceFilter.cs
@@ -3,6 +3,6 @@ namespace SpeechFlowCsharp.AudioProcessing
 {
     public interface IVoiceFilter
     {
-        bool IsHumanVoice(short[] buffer);
+        bool IsHumanVoice(float[] buffer);
     }
 }
